requirements.txt
alembic==0.8.9
coverage==4.1
decorator==4.0.11
psycopg2==2.6.2
pytest-cov==2.2.1
pytest-mock==1.1
pytest-sugar==0.7.1
python-decouple==3.0
SQLAlchemy==1.1.5
coveralls==1.1
Unipath==1.1
